refactor: rename bean to type
