+++ core/src/main/java/org/nervos/ckb/system/type/CkbSystemContract.java
@@ -1,13 +1,14 @@
-package org.nervos.ckb.system.bean;
+package org.nervos.ckb.system.type;
 
 import org.nervos.ckb.methods.type.CellOutPoint;
 
 public class CkbSystemContract {
+
   public String systemScriptCellHash;
   public CellOutPoint systemScriptOutPoint;
 
-  public CkbSystemContract(String systemSriptCellHash, CellOutPoint systemScriptOutPoint) {
-    this.systemScriptCellHash = systemSriptCellHash;
+  public CkbSystemContract(String systemScriptCellHash, CellOutPoint systemScriptOutPoint) {
+    this.systemScriptCellHash = systemScriptCellHash;
     this.systemScriptOutPoint = systemScriptOutPoint;
   }
 }
