+++ core/src/test/java/system/SystemContractTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.nervos.ckb.service.CKBService;
 import org.nervos.ckb.service.HttpService;
 import org.nervos.ckb.system.SystemContract;
-import org.nervos.ckb.system.bean.CkbSystemContract;
+import org.nervos.ckb.system.type.CkbSystemContract;
 import org.nervos.ckb.utils.Network;
 
 public class SystemContractTest {
